refactor: databaseTypes로 데이터베이스 타입이 선언된 파일들의 이름을 통일했다.
